review: refactor, remove useless code and move file to the good dir
