+++ src/docker-compose.yml
@@ -21,173 +21,5 @@ services:
       - "traefik.http.services.traefik.loadbalancer.server.port=8080"
     networks:
       - traefik-global-proxy
-
-
-    
-
-  landy-app6:
-    image: landy-app6:latest
-    container_name: landy-app6
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app6.rule=Host(`landy-app6.local`)"
-      - "traefik.http.routers.landy-app6.entryPoints=websecure"
-      - "traefik.http.routers.landy-app6.tls=true"
-      - "traefik.http.services.landy-app6.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app4:
-    image: landy-app4:latest
-    container_name: landy-app4
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app4.rule=Host(`landy-app4.local`)"
-      - "traefik.http.routers.landy-app4.entryPoints=websecure"
-      - "traefik.http.routers.landy-app4.tls=true"
-      - "traefik.http.services.landy-app4.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app9:
-    image: landy-app9:latest
-    container_name: landy-app9
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app9.rule=Host(`landy-app9.local`)"
-      - "traefik.http.routers.landy-app9.entryPoints=websecure"
-      - "traefik.http.routers.landy-app9.tls=true"
-      - "traefik.http.services.landy-app9.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app1:
-    image: landy-app1:latest
-    container_name: landy-app1
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app1.rule=Host(`landy-app1.local`)"
-      - "traefik.http.routers.landy-app1.entryPoints=websecure"
-      - "traefik.http.routers.landy-app1.tls=true"
-      - "traefik.http.services.landy-app1.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app10:
-    image: landy-app10:latest
-    container_name: landy-app10
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app10.rule=Host(`landy-app10.local`)"
-      - "traefik.http.routers.landy-app10.entryPoints=websecure"
-      - "traefik.http.routers.landy-app10.tls=true"
-      - "traefik.http.services.landy-app10.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app13:
-    image: landy-app13:latest
-    container_name: landy-app13
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app13.rule=Host(`landy-app13.local`)"
-      - "traefik.http.routers.landy-app13.entryPoints=websecure"
-      - "traefik.http.routers.landy-app13.tls=true"
-      - "traefik.http.services.landy-app13.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-        
-
-  landy-app100:
-    image: landy-app100:latest
-    container_name: landy-app100
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app100.rule=Host(`landy-app100.local`)"
-      - "traefik.http.routers.landy-app100.entryPoints=websecure"
-      - "traefik.http.routers.landy-app100.tls=true"
-      - "traefik.http.services.landy-app100.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app1000:
-    image: landy-app1000:latest
-    container_name: landy-app1000
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app1000.rule=Host(`landy-app1000.local`)"
-      - "traefik.http.routers.landy-app1000.entryPoints=websecure"
-      - "traefik.http.routers.landy-app1000.tls=true"
-      - "traefik.http.services.landy-app1000.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-app10000:
-    image: landy-app10000:latest
-    container_name: landy-app10000
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-app10000.rule=Host(`landy-app10000.local`)"
-      - "traefik.http.routers.landy-app10000.entryPoints=websecure"
-      - "traefik.http.routers.landy-app10000.tls=true"
-      - "traefik.http.services.landy-app10000.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy-ap0:
-    image: landy-ap0:latest
-    container_name: landy-ap0
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy-ap0.rule=Host(`landy-ap0.local`)"
-      - "traefik.http.routers.landy-ap0.entryPoints=websecure"
-      - "traefik.http.routers.landy-ap0.tls=true"
-      - "traefik.http.services.landy-ap0.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy2-app:
-    image: landy2-app:latest
-    container_name: landy2-app
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy2-app.rule=Host(`landy2-app.local`)"
-      - "traefik.http.routers.landy2-app.entryPoints=websecure"
-      - "traefik.http.routers.landy2-app.tls=true"
-      - "traefik.http.services.landy2-app.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
-    
-
-  landy22-app:
-    image: landy22-app:latest
-    container_name: landy22-app
-    labels:
-      - "traefik.enable=true"
-      - "traefik.http.routers.landy22-app.rule=Host(`landy22-app.local`)"
-      - "traefik.http.routers.landy22-app.entryPoints=websecure"
-      - "traefik.http.routers.landy22-app.tls=true"
-      - "traefik.http.services.landy22-app.loadbalancer.server.port=3000"
-    networks:
-      - traefik-global-proxy
-     
+
