+++ src/routes.rs
@@ -1,23 +1,12 @@
 use warp::reject;
 use warp::Filter;
-use std::process::Command;
 
-
-use crate::traefik::app_service; 
-use crate::traefik::app_service::docker_compose;
-use crate::traefik::app_service::verif_app;
-
-
-// Le reste de votre code routes.rs
+use crate::services::helpers::traefik_helper::{add_to_deploy, add_to_hosts, verif_app};
 
 use crate::services::helpers::docker_helper::{
-    build_image, create_and_run_container, generate_and_write_dockerfile,
+    build_image, docker_compose, generate_and_write_dockerfile,
 };
 use crate::services::helpers::github_helper::{clone_repo, create_temp_dir, remove_temp_dir};
-use crate::traefik::app_service::add_to_deploy;
-use crate::traefik::app_service::add_to_hosts;
-// use crate::traefik::app_service::add_traefik;
-// use crate::traefik::app_service::create_network;
 
 #[derive(Debug)]
 struct CustomError(String);
@@ -110,14 +99,29 @@ async fn handle_create_app(body: serde_json::Value) -> Result<impl warp::Reply, 
             warp::reject::custom(CustomError(format!("Failed to build Docker image: {}", e)))
         })?;
 
-        add_to_hosts(app_name);
+        add_to_hosts(app_name).map_err(|e| {
+            warp::reject::custom(CustomError(format!(
+                "Failed to add app to hosts file: {}",
+                e
+            )))
+        })?;
 
         if let Ok(1) = verif_app(app_name) {
             println!("L'application {} est déjà déployée.", app_name);
-        } 
-        else {
-            add_to_deploy(app_name, "3000");
-            docker_compose();
+        } else {
+            add_to_deploy(app_name, "3000").map_err(|e| {
+                warp::reject::custom(CustomError(format!(
+                    "Failed to add app to deploy file: {}",
+                    e
+                )))
+            })?;
+
+            docker_compose().map_err(|e| {
+                warp::reject::custom(CustomError(format!(
+                    "Failed to execute docker compose: {}",
+                    e
+                )))
+            })?;
         }
 
         remove_temp_dir(&directory).map_err(|e| {
