+++ src/services/helpers/mod.rs
@@ -1,2 +1,3 @@
 pub mod docker_helper;
 pub mod github_helper;
+pub mod traefik_helper;
