--- src/traefik/mod.rs
@@ -1,1 +1,0 @@
-pub mod app_service;
